requirements.txt
pytest==7.4.3
python-dotenv==1.0.0
requests==2.31.0
openai==1.12.0
openai-agents==0.0.15
streamlit==1.31.1
tiktoken==0.5.2
